main now carries the CI restoration (PR #26), so the -Werror gate
actually reaches this branch's code for the first time. It found four
things this branch had introduced while the gate was down:

- interactive_json.cpp: evidence_type_str() had no case for
  EvidenceType::digital, so a digital evidence token serialised to the UI
  as "unknown". A real bug in the bridge, not a style nit; the switch is
  now exhaustive.
- random_events.cpp: `state` became unused in apply_human_per_tick when
  the per-tick re-announcement was removed. Named out, since the
  parameter is part of the shared handler signature.
- player_loop_harness.h / player_loop_observe.cpp: float members
  accumulated into double totals. Explicit casts; the arithmetic is
  unchanged.

Conflicts were both append-vs-append and are resolved by keeping both
sides: flagged_issues.md (this branch's 2026-09-08 findings plus #26's
2026-09-09 ones) and the tail of population_aging_test.cpp (this
branch's player-clock tests).

  Release + ECONLIFE_WARN_AS_ERROR=ON   builds clean
  clang-format-18 --dry-run --Werror    0 offenders
  ctest -LE emergence                   1899/1899 pass

+++ simulation/tests/integration/history_gen_integration_test.cpp
@@ -90,7 +90,7 @@ TEST_CASE("history gen: founding-seed world bootstraps an economy from zero firm
     auto total_capital = [](const WorldState& w) {
         double t = 0.0;
         for (const auto& n : w.significant_npcs)
-            t += n.capital;
+            t += static_cast<double>(n.capital);
         return t;
     };
     auto solvent = [](const WorldState& w) {
@@ -120,7 +120,7 @@ TEST_CASE("history gen: founding-seed world bootstraps an economy from zero firm
     // capital grows year over year through the run, and a real share of the
     // population is solvent at the end (income reaches people, not just owners).
     WorldState w1 = generate_world_with_history(base_config(42, true), PackageConfig{}, 1, 1);
-    CHECK(total_capital(w) > total_capital(w1));            // growing as it develops
+    CHECK(total_capital(w) > total_capital(w1));                          // growing as it develops
     CHECK(solvent(w) > static_cast<int>(w.significant_npcs.size() / 4));  // broad-based income
 }
